minor fixes and typos

+++ lib/pages/feed_entries/feed_light/form/feed_light_form_bloc.dart
@@ -53,7 +53,7 @@ class FeedLightFormBlocEventCreate extends FeedLightFormBlocEvent {
 
 class FeedLightFormBlocValueChangedEvent extends FeedLightFormBlocEvent {
   final int i;
-  final double value;
+  final int value;
 
   FeedLightFormBlocValueChangedEvent(this.i, this.value);
